Extracted session management logic into a separate class
