+++ src/main/java/io/github/jbellis/brokk/MainProject.java
@@ -13,27 +13,19 @@ import io.github.jbellis.brokk.context.ContextFragment;
 import io.github.jbellis.brokk.context.ContextHistory;
 import io.github.jbellis.brokk.git.GitRepo;
 import io.github.jbellis.brokk.util.AtomicWrites;
-import io.github.jbellis.brokk.util.HistoryIo;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.io.IOException;
-import io.github.jbellis.brokk.util.SerialByKeyExecutor;
 
-import java.io.UncheckedIOException;
 import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.StandardOpenOption;
 import java.util.*;
 import java.util.ArrayList;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import java.util.Set;
 import java.util.LinkedHashSet;
@@ -42,9 +34,7 @@ import java.util.HashSet;
 public final class MainProject extends AbstractProject {
     private static final Logger logger = LogManager.getLogger(MainProject.class); // Separate logger from AbstractProject
 
-    private final ExecutorService sessionExecutor;
-    private final SerialByKeyExecutor sessionExecutorByKey;
-
+    private final SessionManager sessionManager;
     private final Path propertiesFile;
     private final Properties projectProps;
     private final Path styleGuidePath;
@@ -97,8 +87,6 @@ public final class MainProject extends AbstractProject {
     private static final Path PROJECTS_PROPERTIES_PATH = BROKK_CONFIG_DIR.resolve("projects.properties");
     private static final Path GLOBAL_PROPERTIES_PATH = BROKK_CONFIG_DIR.resolve("brokk.properties");
 
-    private final Path sessionsDir;
-
     public enum LlmProxySetting {BROKK, LOCALHOST, STAGING}
 
     private static final String LLM_PROXY_SETTING_KEY = "llmProxySetting";
@@ -117,8 +105,6 @@ public final class MainProject extends AbstractProject {
             - what parts are the trickiest and how could they be simplified
             """.stripIndent();
 
-    private final Map<UUID, SessionInfo> sessionsCache;
-
     public record ProjectPersistentInfo(long lastOpened, List<String> openWorktrees) {
         public ProjectPersistentInfo {
         }
@@ -131,18 +117,11 @@ public final class MainProject extends AbstractProject {
     public MainProject(Path root) {
         super(root); // Initializes this.root and this.repo
 
-        this.sessionExecutor = Executors.newFixedThreadPool(3, r -> {
-            var t = Executors.defaultThreadFactory().newThread(r);
-            t.setDaemon(false);
-            t.setName("session-io-" + t.threadId());
-            return t;
-        });
-        this.sessionExecutorByKey = new SerialByKeyExecutor(sessionExecutor);
-
         this.propertiesFile = this.masterRootPathForConfig.resolve(".brokk").resolve("project.properties");
         this.styleGuidePath = this.masterRootPathForConfig.resolve(".brokk").resolve("style.md");
         this.reviewGuidePath = this.masterRootPathForConfig.resolve(".brokk").resolve("review.md");
-        this.sessionsDir = this.masterRootPathForConfig.resolve(".brokk").resolve("sessions");
+        var sessionsDir = this.masterRootPathForConfig.resolve(".brokk").resolve("sessions");
+        this.sessionManager = new SessionManager(sessionsDir);
         this.mainWorkspacePropertiesPath = this.root.resolve(".brokk").resolve("workspace.properties");
         this.mainWorkspaceProps = new Properties();
 
@@ -225,7 +204,6 @@ public final class MainProject extends AbstractProject {
 
         // Initialize cache and trigger migration/defaulting if necessary
         this.issuesProviderCache = getIssuesProvider();
-        this.sessionsCache = loadSessions();
     }
 
     @Override
@@ -747,119 +725,14 @@ public final class MainProject extends AbstractProject {
         }
     }
 
-    private Path getSessionHistoryPath(UUID sessionId) {
-        return sessionsDir.resolve(sessionId.toString() + ".zip");
-    }
-
-    private Optional<SessionInfo> readSessionInfoFromZip(Path zipPath) {
-        if (!Files.exists(zipPath)) return Optional.empty();
-        try (var fs = FileSystems.newFileSystem(zipPath, Map.of())) {
-            Path manifestPath = fs.getPath("manifest.json");
-            if (Files.exists(manifestPath)) {
-                String json = Files.readString(manifestPath);
-                return Optional.of(objectMapper.readValue(json, SessionInfo.class));
-            }
-        } catch (IOException e) {
-            logger.warn("Error reading manifest.json from {}: {}", zipPath.getFileName(), e.getMessage());
-        }
-        return Optional.empty();
-    }
-
-    private void writeSessionInfoToZip(Path zipPath, SessionInfo sessionInfo) throws IOException {
-        try (var fs = FileSystems.newFileSystem(zipPath, Map.of("create", Files.notExists(zipPath) ? "true" : "false"))) {
-            Path manifestPath = fs.getPath("manifest.json");
-            String json = objectMapper.writeValueAsString(sessionInfo);
-            Files.writeString(manifestPath, json, StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING);
-        } catch (IOException e) {
-            logger.error("Error writing manifest.json to {}: {}", zipPath.getFileName(), e.getMessage());
-            throw e;
-        }
-    }
-
     @Override
     public void saveHistory(ContextHistory ch, UUID sessionId) {
-        SessionInfo infoToSave = null;
-        SessionInfo currentInfo = sessionsCache.get(sessionId);
-        if (currentInfo != null) {
-            infoToSave = new SessionInfo(currentInfo.id(), currentInfo.name(), currentInfo.created(), System.currentTimeMillis());
-            sessionsCache.put(sessionId, infoToSave); // Update cache before async task
-        } else {
-            logger.warn("Session ID {} not found in cache. History content will be saved, but manifest cannot be updated.", sessionId);
-        }
-
-        final SessionInfo finalInfoToSave = infoToSave;
-        sessionExecutorByKey.submit(sessionId.toString(), () -> {
-            try {
-                Path sessionHistoryPath = getSessionHistoryPath(sessionId);
-                HistoryIo.writeZip(ch, sessionHistoryPath);
-                if (finalInfoToSave != null) {
-                    writeSessionInfoToZip(sessionHistoryPath, finalInfoToSave);
-                }
-            } catch (IOException e) {
-                logger.error("Error saving context history or updating manifest for session {}: {}", sessionId, e.getMessage());
-            }
-        });
+        sessionManager.saveHistory(ch, sessionId);
     }
 
     @Override
     public @Nullable ContextHistory loadHistory(UUID sessionId, IContextManager contextManager) {
-        var future = sessionExecutorByKey.submit(sessionId.toString(), () -> {
-            try {
-                var sessionHistoryPath = getSessionHistoryPath(sessionId);
-                ContextHistory ch = HistoryIo.readZip(sessionHistoryPath, contextManager);
-                if (ch == null) {
-                    return null;
-                }
-                // Resetting nextId based on loaded fragments.
-                // Only consider numeric IDs for dynamic fragments.
-                // Hashes will not parse to int and will be skipped by this logic.
-                int maxNumericId = 0;
-                for (Context ctx : ch.getHistory()) {
-                    for (ContextFragment fragment : ctx.allFragments().toList()) {
-                        try {
-                            maxNumericId = Math.max(maxNumericId, Integer.parseInt(fragment.id()));
-                        } catch (NumberFormatException e) {
-                            // Ignore non-numeric IDs (hashes)
-                        }
-                    }
-                    for (TaskEntry taskEntry : ctx.getTaskHistory()) {
-                        if (taskEntry.log() != null) {
-                            try {
-                                // TaskFragment IDs are hashes, so this typically won't contribute to maxNumericId.
-                                // If some TaskFragments had numeric IDs historically, this would catch them.
-                                maxNumericId = Math.max(maxNumericId, Integer.parseInt(taskEntry.log().id()));
-                            } catch (NumberFormatException e) {
-                                // Ignore non-numeric IDs
-                            }
-                        }
-                    }
-                }
-                // ContextFragment.nextId is an AtomicInteger, its value is the *next* ID to be assigned.
-                // If maxNumericId found is, say, 10, nextId should be set to 10 so that getAndIncrement() yields 11.
-                // If setNextId ensures nextId will be value+1, then passing maxNumericId is correct.
-                // Current ContextFragment.setNextId: if (value >= nextId.get()) { nextId.set(value); }
-                // Then nextId.getAndIncrement() will use `value` and then increment it.
-                // So we should set it to maxNumericId found.
-                if (maxNumericId > 0) { // Only set if we found any numeric IDs
-                     ContextFragment.setMinimumId(maxNumericId + 1);
-                     logger.debug("Restored dynamic fragment ID counter based on max numeric ID: {}", maxNumericId);
-                }
-                return ch;
-            } catch (IOException e) {
-                logger.error("Error loading context history for session {}: {}", sessionId, e.getMessage());
-                return null;
-            }
-        });
-
-        try {
-            return future.get();
-        } catch (InterruptedException | ExecutionException e) {
-            logger.warn("Error waiting for session history to load for session {}: {}", sessionId, e);
-            if (e instanceof InterruptedException) {
-                Thread.currentThread().interrupt();
-            }
-            return null;
-        }
+        return sessionManager.loadHistory(sessionId, contextManager);
     }
 
     public static LlmProxySetting getProxySetting() {
@@ -1352,191 +1225,38 @@ public final class MainProject extends AbstractProject {
     }
     
     public static Optional<String> getActiveSessionTitle(Path worktreeRoot) {
-        var wsPropsPath = worktreeRoot.resolve(".brokk").resolve("workspace.properties");
-        if (!Files.exists(wsPropsPath)) {
-            return Optional.empty();
-        }
-        var props = new Properties();
-        try (var reader = Files.newBufferedReader(wsPropsPath)) {
-            props.load(reader);
-        } catch (IOException e) {
-            logger.warn("Error reading workspace properties at {}: {}", wsPropsPath, e.getMessage());
-            return Optional.empty();
-        }
-        String sessionIdStr = props.getProperty("lastActiveSession");
-        if (sessionIdStr == null || sessionIdStr.isBlank()) {
-            return Optional.empty();
-        }
-        UUID sessionId;
-        try {
-            sessionId = UUID.fromString(sessionIdStr.trim());
-        } catch (IllegalArgumentException e) {
-            logger.warn("Invalid session UUID '{}' in workspace properties at {}", sessionIdStr, wsPropsPath);
-            return Optional.empty();
-        }
-        Path masterRootPath;
-        if (GitRepo.hasGitRepo(worktreeRoot)) {
-            try (var tempRepo = new GitRepo(worktreeRoot)) {
-                masterRootPath = tempRepo.getGitTopLevel();
-            } catch (Exception e) {
-                logger.warn("Error determining git top level for {}: {}", worktreeRoot, e.getMessage());
-                return Optional.empty();
-            }
-        } else {
-            masterRootPath = worktreeRoot;
-        }
-        Path sessionZip = masterRootPath.resolve(".brokk").resolve("sessions").resolve(sessionId + ".zip");
-        if (!Files.exists(sessionZip)) {
-            logger.debug("Session zip not found at {} for session ID {}", sessionZip, sessionId);
-            return Optional.empty();
-        }
-        try (var fs = FileSystems.newFileSystem(sessionZip, Map.of())) {
-            Path manifestPath = fs.getPath("manifest.json");
-            if (Files.exists(manifestPath)) {
-                String json = Files.readString(manifestPath);
-                var sessionInfo = objectMapper.readValue(json, SessionInfo.class);
-                return Optional.of(sessionInfo.name());
-            }
-        } catch (IOException e) {
-            logger.warn("Error reading session manifest from {}: {}", sessionZip.getFileName(), e.getMessage());
-        }
-        return Optional.empty();
-    }
-
-
-    private Map<UUID, SessionInfo> loadSessions() {
-        var sessions = new ConcurrentHashMap<UUID, SessionInfo>();
-        try {
-            Files.createDirectories(sessionsDir);
-            try (var stream = Files.list(sessionsDir)) {
-                stream.filter(path -> path.toString().endsWith(".zip"))
-                        .forEach(zipPath -> readSessionInfoFromZip(zipPath).ifPresent(sessionInfo -> {
-                            sessions.put(sessionInfo.id(), sessionInfo);
-                        }));
-            }
-        } catch (IOException e) {
-            logger.error("Error listing session zip files in {}: {}", sessionsDir, e.getMessage());
-        }
-        return sessions;
+        return SessionManager.getActiveSessionTitle(worktreeRoot);
     }
 
     @Override
     public List<SessionInfo> listSessions() {
-        var sessions = new ArrayList<>(sessionsCache.values());
-        sessions.sort(Comparator.comparingLong(SessionInfo::modified).reversed());
-        return sessions;
+        return sessionManager.listSessions();
     }
 
     @Override
     public SessionInfo newSession(String name) {
-        var sessionId = UUID.randomUUID();
-        var currentTime = System.currentTimeMillis();
-        var newSessionInfo = new SessionInfo(sessionId, name, currentTime, currentTime);
-        sessionsCache.put(sessionId, newSessionInfo);
-
-        sessionExecutorByKey.submit(sessionId.toString(), () -> {
-            Path sessionHistoryPath = getSessionHistoryPath(sessionId);
-            try {
-                Files.createDirectories(sessionHistoryPath.getParent());
-                // 1. Create the zip with empty history first. This ensures the zip file exists.
-                var emptyHistory = new ContextHistory(Context.EMPTY);
-                HistoryIo.writeZip(emptyHistory, sessionHistoryPath);
-
-                // 2. Now add/update manifest.json to the existing zip.
-                writeSessionInfoToZip(sessionHistoryPath, newSessionInfo); // Should use create="false" as zip exists.
-                logger.info("Created new session {} ({}) with manifest and empty history.", name, sessionId);
-            } catch (IOException e) {
-                logger.error("Error creating new session files for {} ({}): {}", name, sessionId, e.getMessage());
-                throw new UncheckedIOException("Failed to create new session " + name, e);
-            }
-        });
-        return newSessionInfo;
+        return sessionManager.newSession(name);
     }
 
     @Override
     public void renameSession(UUID sessionId, String newName) {
-        SessionInfo oldInfo = sessionsCache.get(sessionId);
-        if (oldInfo != null) {
-            var updatedInfo = new SessionInfo(oldInfo.id(), newName, oldInfo.created(), System.currentTimeMillis());
-            sessionsCache.put(sessionId, updatedInfo);
-            sessionExecutorByKey.submit(sessionId.toString(), () -> {
-                try {
-                    Path sessionHistoryPath = getSessionHistoryPath(sessionId);
-                    writeSessionInfoToZip(sessionHistoryPath, updatedInfo);
-                    logger.info("Renamed session {} to '{}'", sessionId, newName);
-                } catch (IOException e) {
-                    logger.error("Error writing updated manifest for renamed session {}: {}", sessionId, e.getMessage());
-                }
-            });
-        } else {
-            logger.warn("Session ID {} not found in cache, cannot rename.", sessionId);
-        }
+        sessionManager.renameSession(sessionId, newName);
     }
 
     @Override
     public void deleteSession(UUID sessionId) {
-        sessionsCache.remove(sessionId);
-        sessionExecutorByKey.submit(sessionId.toString(), () -> {
-            Path historyZipPath = getSessionHistoryPath(sessionId);
-            try {
-                boolean deleted = Files.deleteIfExists(historyZipPath);
-                if (deleted) {
-                    logger.info("Deleted session zip: {}", historyZipPath.getFileName());
-                } else {
-                    logger.warn("Session zip {} not found for deletion, or already deleted.", historyZipPath.getFileName());
-                }
-            } catch (IOException e) {
-                logger.error("Error deleting history zip for session {}: {}", sessionId, e.getMessage());
-            }
-        });
+        sessionManager.deleteSession(sessionId);
     }
 
     @Override
     public SessionInfo copySession(UUID originalSessionId, String newSessionName) throws IOException {
-        var newSessionId = UUID.randomUUID();
-        var currentTime = System.currentTimeMillis();
-        var newSessionInfo = new SessionInfo(newSessionId, newSessionName, currentTime, currentTime);
-        sessionsCache.put(newSessionId, newSessionInfo);
-
-        var keys = Set.of(originalSessionId.toString(), newSessionId.toString());
-        sessionExecutorByKey.submit(keys, () -> {
-            try {
-                Path originalHistoryPath = getSessionHistoryPath(originalSessionId);
-                if (!Files.exists(originalHistoryPath)) {
-                    throw new IOException("Original session %s not found, cannot copy".formatted(originalHistoryPath.getFileName()));
-                }
-                
-                Path newHistoryPath = getSessionHistoryPath(newSessionId);
-                Files.createDirectories(newHistoryPath.getParent());
-                Files.copy(originalHistoryPath, newHistoryPath);
-                logger.info("Copied session zip {} to {}", originalHistoryPath.getFileName(), newHistoryPath.getFileName());
-
-                writeSessionInfoToZip(newHistoryPath, newSessionInfo);
-                logger.info("Updated manifest.json in new session zip {} for session ID {}", newHistoryPath.getFileName(), newSessionId);
-            } catch (Exception e) {
-                logger.error("Failed to copy session from {} to new session {}: {}", originalSessionId, newSessionName, e.getMessage(), e);
-                if (e instanceof InterruptedException) {
-                    Thread.currentThread().interrupt();
-                }
-                throw new RuntimeException("Failed to copy session " + originalSessionId, e);
-            }
-        });
-        return newSessionInfo;
+        return sessionManager.copySession(originalSessionId, newSessionName);
     }
 
     @Override
     public void close() {
         super.close();
-        sessionExecutor.shutdown();
-        try {
-            if (!sessionExecutor.awaitTermination(30, TimeUnit.SECONDS)) {
-                logger.warn("Session IO tasks did not finish in 30 seconds, forcing shutdown.");
-                sessionExecutor.shutdownNow();
-            }
-        } catch (InterruptedException e) {
-            sessionExecutor.shutdownNow();
-            Thread.currentThread().interrupt();
-        }
+        sessionManager.close();
     }
 
     public Path getWorktreeStoragePath() {
